Attack Player + Follow
completed sprint to attack player + the monster can now follow whatever destination we give it for example if we put a function on a door it will run towards that door etc. also added animations for attacking and running. if the monster is very close to the player it will attack and deal damage. I also put a test function on PlayerMovement script to trigger following which can be deleted later once everything is tested.